[CSL-1131] Split node/Main.hs on Secrets.hs and Params.hs
