+++ cardano-sl.cabal
@@ -563,6 +563,8 @@ executable cardano-node
   hs-source-dirs:      src/node
   main-is:             Main.hs
   other-modules:       NodeOptions
+                     , Params
+                     , Secrets
   build-depends:       base
                      , binary
                      , bytestring
